feat: move react-script -> vite
