+++ api/dev.js
@@ -1,0 +1,5 @@
+const { setupClient } = require('vite-plugin-blocklet');
+
+const { app } = require('./index');
+
+setupClient(app);
