+++ package.json
@@ -7,15 +7,10 @@
   "scripts": {
     "lint": "eslint src api",
     "clean": "rm -rf build",
-    "start": "npm-run-all --parallel start:*",
-    "start:client": "NODE_ENV=development react-app-rewired start",
-    "start:server": "nodemon api/index.js -w api",
-    "bundle": "yarn clean && npm run bundle:client && npm run bundle:server",
-    "bundle:client": "PUBLIC_URL=\".\" REACT_APP_API_PREFIX=\"\" REACT_APP_APP_NAME=\"Profile Demo\" yarn build:client",
-    "build:client": "react-scripts build && rm -f build/**/**/*.{js,css}.map",
-    "bundle:server": "blocklet bundle --create-release",
-    "test": "node tools/jest.js",
-    "release": "npm run clean && npm run bundle",
+    "start": "NODE_ENV=development nodemon api/dev.js -w api",
+    "bundle": "npm run clean && npm run bundle:client && npm run bundle:server",
+    "bundle:client": "vite build",
+    "bundle:server": "bn bundle --create-release",
     "coverage": "npm run test -- --coverage"
   },
   "dependencies": {
@@ -29,9 +24,7 @@
     "@mui/lab": "^5.0.0-alpha.104",
     "@mui/material": "^5.10.10",
     "@mui/styles": "^5.10.10",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
+    "@vitejs/plugin-react": "^3.1.0",
     "axios": "^0.21.4",
     "body-parser": "^1.20.1",
     "compression": "^1.7.4",
@@ -47,12 +40,12 @@
     "morgan": "^1.10.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "4.0.2"
+    "vite": "^4.1.4",
+    "vite-plugin-blocklet": "^0.5.18"
   },
   "eslintConfig": {
     "extends": [
-      "react-app",
-      "react-app/jest"
+      "react-app"
     ]
   },
   "browserslist": {
@@ -68,10 +61,7 @@
     ]
   },
   "devDependencies": {
-    "http-proxy-middleware": "^2.0.6",
-    "jest": "^29.2.1",
     "nodemon": "^2.0.20",
-    "npm-run-all": "^4.1.5",
-    "react-app-rewired": "^2.2.1"
+    "npm-run-all": "^4.1.5"
   }
 }
