+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import App from './app';
 
-const root = createRoot(document.getElementById('root'));
+const root = createRoot(document.getElementById('app'));
 root.render(<App />);
